apps/api/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
pyjwt[crypto]
httpx

# Database
asyncpg
sqlalchemy
psycopg2-binary
psycopg[binary]  # Added for LlamaIndex Chat Store

# AI / RAG
llama-index>=0.11.0
llama-index-core
llama-index-vector-stores-postgres
llama-index-storage-chat-store-postgres
llama-index-embeddings-openai
llama-index-llms-openai
openai
